+++ src/api.rs
@@ -161,8 +161,7 @@ mod tests {
 
     use futures::Future;
 
-    use crate::storage::MockStorage;
-    use crate::testing_utils::{gen_dummy_hash,gen_challenge_state};
+    use crate::testing_utils::{gen_dummy_hash,gen_challenge_state,MockStorage};
 
     #[test]
     fn get_request_test() {
