+++ src/challenger.rs
@@ -101,6 +101,7 @@ pub fn run_challenge_request<T: Service, K: ClientChain, D: Storage>(
     loop {
         let challenge_height = service.get_blockheight()?;
         info! {"service chain height: {}", challenge_height}
+        println!("prev: {} challfreq: {}", prev_challenge_height, challenge_frequency);
         if (request.end_blockheight as u64) < challenge_height {
             break;
         } else if (challenge_height - prev_challenge_height) < challenge_frequency {
@@ -208,11 +209,8 @@ mod tests {
 
     use std::sync::mpsc::{channel, Receiver, Sender};
 
-    use crate::clientchain::MockClientChain;
     use crate::error::Error;
-    use crate::service::MockService;
-    use crate::storage::MockStorage;
-    use crate::testing_utils::gen_dummy_hash;
+    use crate::testing_utils::{gen_dummy_hash,MockClientChain,MockService,MockStorage};
 
     #[test]
     fn verify_challenge_test() {
