+++ src/clientchain.rs
@@ -2,10 +2,9 @@
 //!
 //! Client chain interface and implementations
 
-use std::cell::RefCell;
 use std::collections::HashMap;
 
-use bitcoin_hashes::{hex::FromHex, sha256d, Hash};
+use bitcoin_hashes::{hex::FromHex, sha256d};
 use ocean_rpc::{json, RpcApi};
 
 use crate::config::ClientChainConfig;
@@ -128,50 +127,5 @@ impl<'a> ClientChain for RpcClientChain<'a> {
             Err(e) => warn!("verify challenge error{}", e),
         }
         Ok(false)
-    }
-}
-
-/// Mock implementation of ClientChain using some mock logic for testing
-pub struct MockClientChain {
-    /// Flag that when set returns error on all inherited methods that return
-    /// Result
-    pub return_err: bool,
-    /// Flag that when set returns false on all inherited methods that return
-    /// bool
-    pub return_false: bool,
-    /// Mock client chain blockheight
-    pub height: RefCell<u64>,
-}
-
-impl MockClientChain {
-    /// Create a MockClientChain with all flags turned off by default
-    pub fn new() -> Self {
-        MockClientChain {
-            return_err: false,
-            return_false: false,
-            height: RefCell::new(0),
-        }
-    }
-}
-
-impl ClientChain for MockClientChain {
-    /// Send challenge transaction to client chain
-    fn send_challenge(&self) -> Result<sha256d::Hash> {
-        if self.return_err {
-            return Err(Error::from(CError::Generic("send_challenge failed".to_owned())));
-        }
-        // Use height to generate mock challenge hash
-        Ok(sha256d::Hash::from_slice(&[(*self.height.borrow() % 16) as u8; 32])?)
-    }
-
-    /// Verify challenge transaction has been included in the chain
-    fn verify_challenge(&self, _txid: &sha256d::Hash) -> Result<bool> {
-        if self.return_err {
-            return Err(Error::from(CError::Generic("verify_challenge failed".to_owned())));
-        }
-        if self.return_false {
-            return Ok(false);
-        }
-        Ok(true)
     }
 }
